% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as.stream.R
\name{as.stream}
\alias{as.stream}
\alias{print.stream}
\title{Converts a character vector to an object of class "stream"}
\usage{
as.stream(obj)

\method{print}{stream}(x, ...)
}
\arguments{
\item{obj}{The character vector to be converted to a "stream" object}

\item{x}{The "stream" object to be printed.}

\item{\dots}{Ignored}
}
\value{
The same character vector, only it now inherits from class "stream"
}
\description{
Converts a character vector to an object of class "stream"
}
\section{Methods (by generic)}{
\itemize{
\item \code{print}: Prints an object of class "stream" by removing quotes from the strings and printing each element
of the character vector on its own line.
}}

\examples{
aStream <- c("string1", "string2", "string3")
aStream

# Note how it prints after we convert it to a stream object:
as.stream(aStream)
}
\author{
Landon Sego
}
