% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rsed-package.R
\docType{package}
\name{rsed-package}
\alias{rsed-package}
\alias{rsed}
\title{rsed}
\description{
rsed:  Stream editing in R. Tools for manipulating text files with insertions, replacements, deletions, substitutions, and comments.
}
\details{
\tabular{ll}{
Package: \tab rsed\cr
Type: \tab Package\cr
Version: \tab 0.1.1\cr
Date: \tab 2017-11-18\cr
License: \tab BSD_3_clause + file LICENSE\cr
}
}
\examples{
help(package = rsed)
}
\author{
Landon Sego

Maintainer: Landon Sego <LHS@byu.net>
}
\keyword{package}
