% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rsed-package.R
\docType{package}
\name{rsed}
\alias{rsed}
\alias{rsed-package}
\title{rsed}
\description{
rsed:  Stream editing in R. Tools for manipulating text files with insertions, replacements, deletions, substitutions, and comments.
}
\details{
\tabular{ll}{
Package: \tab rsed\cr
Type: \tab Package\cr
Version: \tab 0.1.0\cr
Date: \tab 2016-03-25\cr
License: \tab LICENSE\cr
}
}
\examples{
help(package = rsed)
}
\author{
Landon Sego

Maintainer: Landon Sego <Landon.Sego@pnnl.gov>
}
\keyword{package}

